src/requirements.txt
discord.py==1.5.1
PyJWT==2.0.0
PyYAML==5.3.1
psycopg2==2.8.6
SQLAlchemy==1.3.22